requirements.txt
fastapi
uvicorn
supabase==1.0.3
python-dotenv
bcrypt
PyJWT
python-jose
pydantic